requirements.txt
langchain
openai
streamlit
streamlit-extras
faiss-cpu
PyPDF2
python-dotenv
tiktoken